chore: dedupe the wgsl preambles
